Remove dead code and add some doc

+++ RangeWeaponOverhaul.cs
@@ -22,13 +22,15 @@ namespace BossRush
         /// </summary>
         public static float SpreadModify;
         /// <summary>
-        /// This is use to set how many projectile you can shoot in actual weapon class
-        /// Do not use NumOfProjectile = 0 as it will make gun unable to shoot
+        /// This is use to set how many projectile you can shoot in actual weapon class<br/>
+        /// Do not use NumOfProjectile = 0 as it will make gun unable to shoot<br/>
+        /// Use this when making a range modder weapon
         /// </summary>
         public static int NumOfProjectile = 1;
         /// <summary>
         /// This is to modify the amount of projectiles you gonna shoot from weapon that got assigned with NumOfProjectile<br/>
-        /// This is the safe way to modify the amount of projectile you can modify, do keep in mind to not go below -1 as it is very dangerous
+        /// This is the safe way to modify the amount of projectile you can modify, do keep in mind to not go below -1 as it is very dangerous<br/>
+        /// Use this if you are making a accessory
         /// </summary>
         public static float NumOfProModify = 0;
         /// <summary>
@@ -130,9 +132,6 @@ namespace BossRush
             /// <param name="SpreadAmount">Rotation radius</param>
             /// <param name="AdditionalSpread">Addition X and Y modifier</param>
             /// <param name="AdditionalMultiplier">Multiplier for final speed change</param>
-            /// <param name="ItemISaShotgun">
-            /// Set true if the Item is a shotgun to make it don't change the angle it aim at <br/>
-            /// Set false if the Item is not a shotgun to make it emulate recoil<br/>
             /// </param>
             public void GlobalRandomSpreadFiring(
                 Player player,
@@ -147,7 +146,7 @@ namespace BossRush
                 float AdditionalSpread = 0,
                 float AdditionalMultiplier = 1)
             {
-                float ProjectileAmount = RangeWeaponOverhaul.ModifiedProjAmount(NumOfProjectile);
+                int ProjectileAmount = (int)RangeWeaponOverhaul.ModifiedProjAmount(NumOfProjectile);
                 if (ProjectileAmount == 1)
                 {
                     velocity = velocity.RotateRandom(SpreadAmount).RandomSpread(AdditionalSpread, AdditionalMultiplier);
